🚚 rename [#38] SVG icon 컴포넌트 파일 이름 변경
